working on project
css working

+++ src/components/SidebarComponents/DeleteEmployeeButton.js
@@ -1,9 +1,19 @@
 import React from "react";
 
 const DeleteEmployeeButton = ({ employeeId, onDelete }) => {
+  const token = localStorage.getItem("authToken"); // Pobranie tokena z localStorage
+
   const handleDelete = () => {
-    fetch(`$/Employees/delete/${employeeId}`, {
+    if (!token) {
+      console.error("Brak tokena autoryzacyjnego. Zaloguj się ponownie.");
+      return;
+    }
+
+    fetch(`/Employees/delete/${employeeId}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${token}`,
+      },
     })
       .then((response) => {
         if (!response.ok) {
